complete version 5
change the version control from "Type" to "version x", delete all old "Type" excluding TypeII for test use. The changes of version 5 are:
1, utilise pointer and reference;
2, include area unit in likelihood function;
3, change "density" to a vector to avoid calculating twice.
